REFACT: refactor the test
